refactor: move java stream events classes to new directory
