+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/model/stream/events/StreamRoomDeletedEventData.java
@@ -1,4 +1,4 @@
-package com.simplito.java.privmx_endpoint.model.events.stream;
+package com.simplito.java.privmx_endpoint.model.stream.events;
 
 public class StreamRoomDeletedEventData {
     public String streamRoomId;
