+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/model/stream/events/StreamUnpublishedEventData.java
@@ -1,4 +1,4 @@
-package com.simplito.java.privmx_endpoint.model.events.stream;
+package com.simplito.java.privmx_endpoint.model.stream.events;
 
 public class StreamUnpublishedEventData {
     public String streamRoomId;
