+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/model/stream/events/StreamsUpdatedData.java
@@ -1,4 +1,4 @@
-package com.simplito.java.privmx_endpoint.model.events.stream;
+package com.simplito.java.privmx_endpoint.model.stream.events;
 
 
 import java.util.List;
